requirements_gpu.txt
fastapi
uvicorn
pydantic
pydantic-settings
aiogram
aiohttp
python-dotenv
openai
torch
diffusers
transformers
accelerate
pillow
moviepy
pydub
ffmpeg-python
pyngrok
TTS
librosa
opencv-python
numpy
